rename lightcone files.
